+++ fml-vscode-extension/vscode-extension/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.0 (2026-08-08)
 * Fix issue with FML validation of properties with base type definitions like `Resource` to permit `Observation` and `Patientm such as found in Bundle.entry.resource
+* Tooltip enhancements for groups/group invocations
 
 ## 0.9.0 (2026-08-07)
 * element autocomplete bug fixes